remove the precentage in stores details discount

+++ app/src/main/java/com/dreamcard/app/view/activity/StoreDetailsActivity.java
@@ -205,7 +205,7 @@ public class StoreDetailsActivity extends Activity implements View.OnClickListen
 
         txtBusinessName.setText(bean.getStoreName());
         if (bean.getDiscountPrecentage() > 0.0) {
-            txtAnnualDiscount.setText(Double.toString(bean.getDiscountPrecentage()) + "%");
+            txtAnnualDiscount.setText(Integer.toString((int)bean.getDiscountPrecentage()) + "%");
             txtAnnualDiscount.setVisibility(View.VISIBLE);
             ratingPnl.setVisibility(View.VISIBLE);
         }
